Improve range reading support
